+++ tests/064_f5Client_qkview.int.tests.ts
@@ -22,7 +22,7 @@ import { AuthTokenReqBody } from '../src/bigip/bigipModels';
 import { iControlEndpoints } from '../src/constants';
 
 
-import { deviceInfoIPv6 } from './artifacts/f5_device_atc_infos';
+import { deviceInfoIPv6 } from '../src/utils/f5_device_atc_infos';
 
 
 let f5Client: F5Client;