+++ tests/066_f5Client_atcMgmt.int.tests.ts
@@ -21,7 +21,7 @@ import { getFakeToken } from '../src/utils/testingUtils';
 import { AuthTokenReqBody } from '../src/bigip/bigipModels';
 import { iControlEndpoints } from '../src/constants';
 
-import { deviceInfoIPv6 } from './artifacts/f5_device_atc_infos';
+import { deviceInfoIPv6 } from '../src/utils/f5_device_atc_infos';
 import { isObject } from '../src/utils/misc';